Extract class filename conversion into own class
