+++ Classes/Mw/Metamorph/Step/CreateClasses.php
@@ -6,6 +6,7 @@ use Mw\Metamorph\Domain\Model\MorphConfiguration;
 use Mw\Metamorph\Domain\Model\State\ClassMapping;
 use Mw\Metamorph\Domain\Repository\MorphConfigurationRepository;
 use Mw\Metamorph\Domain\Service\MorphExecutionState;
+use Mw\Metamorph\Step\ClassNameConversion\ClassToFilenameConversion;
 use Mw\Metamorph\Transformation\AbstractTransformation;
 use Mw\Metamorph\Transformation\Progressible;
 use Mw\Metamorph\Transformation\ProgressibleTrait;
@@ -29,6 +30,12 @@ class CreateClasses extends AbstractTransformation implements Progressible {
 	 */
 	protected $morphRepository;
 
+	/**
+	 * @var ClassToFilenameConversion
+	 * @Flow\Inject
+	 */
+	protected $classFilenameConversion;
+
 	public function execute(MorphConfiguration $configuration, MorphExecutionState $state) {
 		$classMappingContainer = $configuration->getClassMappingContainer();
 		$packageClassCount     = [];
@@ -45,7 +52,7 @@ class CreateClasses extends AbstractTransformation implements Progressible {
 			$newClassName = new String($classMapping->getNewClassName());
 
 			$relativeFilename = $newClassName->replace('\\', '/')->append('.php');
-			$absoluteFilename = $this->getAbsoluteFilename($relativeFilename, $package);
+			$absoluteFilename = $this->classFilenameConversion->getAbsoluteFilename($relativeFilename, $package);
 
 			Files::createDirectoryRecursively(dirname($absoluteFilename));
 
@@ -65,44 +72,6 @@ class CreateClasses extends AbstractTransformation implements Progressible {
 
 		foreach ($packageClassCount as $package => $count) {
 			$this->log('<comment>%d</comment> classes written to package <comment>%s</comment>.', [$count, $package]);
-		}
-	}
-
-	/**
-	 * Determines if a class contains a test case.
-	 *
-	 * @param \Helmich\Scalars\Types\String $relativeFilename The relative class file name.
-	 * @return bool TRUE if the class contains a test case, otherwise FALSE.
-	 */
-	private function isClassTestCase(String $relativeFilename) {
-		return $relativeFilename->strip('/')->contains('Tests/') || $relativeFilename->endsWidth('Test.php');
-	}
-
-	/**
-	 * Gets the absolute target filename for a class file.
-	 *
-	 * @param \Helmich\Scalars\Types\String $relativeFilename The relative class file name (auto-derived from class name).
-	 * @param PackageInterface              $package          The target package.
-	 * @return string The target filename.
-	 */
-	private function getAbsoluteFilename(String $relativeFilename, PackageInterface $package) {
-		if (FALSE === $this->isClassTestCase($relativeFilename)) {
-			return (new String($package->getClassesPath()))
-				->stripRight('/')
-				->append('/')
-				->append($relativeFilename)
-				->toPrimitive();
-		} else {
-			return (new String(''))
-				->append($package->getPackagePath())
-				->stripRight('/')
-				->append('/Tests/Unit/')
-				->append(
-					$relativeFilename
-						->replace('Tests/', '')
-						->replace((new String($package->getPackageKey()))->replace('.', '/')->append('/'), '')
-				)
-				->toPrimitive();
 		}
 	}
 }
